Working selected character and market selection move from [type] to __layout
