+++ src/routes/breakdown/[type].svelte
@@ -38,70 +38,12 @@
 
         selectableTypes = selectableTypeIDs.filter(id=>$Universe.types[id]!==undefined).map(id=>$Universe.types[id]);
     }
-
-    $: blueprints = CharacterBlueprints[ $currentCharacter ];
-
-    const TradeHubs = [
-        60003760,   // Jita IV - Moon 4 - Caldari Navy Assembly Plant
-        1028858195912,   // Perimeter Tranquility Trading Tower
-        // Currently only Jita/Perimeter is supported because market region is hard-coded to The Forge
-        
-        // 60008494,   // Amarr VIII (Oris) - Emperor Family Academy
-        // 60004588,   // Rens VI - Moon 8 - Brutor Tribe Treasury
-        // 60011866,   // Dodixie IX - Moon 20 - Federation Navy Assembly Plant
-        // 60005686,   // Hek VIII - Moon 12 - Boundless Creation Factory
-    ]
-
-    // Set the locations context for all location selectors below
-    let _unsubscribes: Array<Unsubscriber>=[];
-    let locations = writable({});
-    $: if(mounted) {    // Ensure that subscribes run only if the component is mounted
-        _unsubscribes.forEach(u=>u());
-        _unsubscribes = [];
-
-        let _locations = {};
-        locations.set(_locations);
-
-        $blueprints?.forEach(b=>{ 
-            _locations[b.location_id] = null;
-        });
-
-        TradeHubs.forEach(location_id=>_locations[location_id] = null)
-
-        let ids = Object.keys(_locations);
-        ids.forEach(location_id=>{
-            let unsubscribe = GetLocationStore(parseInt(location_id)).subscribe(value=>{
-                _locations[location_id] = value;
-                locations.set(_locations);
-            });
-
-            if(typeof unsubscribe == 'function') _unsubscribes.push(unsubscribe);
-
-        })
-    }
-
-    let mounted = false;
-    onMount(()=>{mounted = true})
-
-    onDestroy(()=>{_unsubscribes.forEach(u=>u&&u())})
-
-    setContext('locations', locations);
-
-    let marketFilterLocation = writable<Location_Id>(null);
-    setContext('marketFilterLocation', marketFilterLocation);
-
-    let currentCharacter = writable<Character_Id>(null);
-    setContext('currentCharacter', currentCharacter);
 </script>
 
 
 <svelte:head>
 	<title>{($Universe?.types && $Universe?.types[selectedTypeId]) ? `${$Universe?.types[selectedTypeId]?.name} - ` : "" }EVE Online Visual Industry Calculator</title>
 </svelte:head>
-
-
-<CharacterSelector bind:value={$currentCharacter} /><br/>
-Filter market <LocationSelector allowUnselected bind:value={$marketFilterLocation} />
 
 
 <TypeSelector {selectedTypeId} on:change={event=>{goto(`${event.detail}`, {keepfocus:true})}} {selectableTypes} />
